moved exercise set views history to a tabpane along with the sets

+++ src/main/java/com/mycompany/controllers/CategoryStatisticsController.java
@@ -288,7 +288,7 @@ public class CategoryStatisticsController {
             addPieChartData(chart, "Other", totalSetsNotTracked);
         }
         
-        chart.setTitle(title + " (" + chart.getTotal() + ")");
+        chart.setTitle(title + " (" + ((int) chart.getTotal()) + ")");
     }
     
     private void addPieChartData(PieChart pChart, String name, int value) {
